+++ HISTORY.rst
@@ -6,6 +6,13 @@ History
 .. to_doc
 
 ---------------------
+0.1.1 (2016-11-17)
+---------------------
+* Make python2/3 compatible
+* Use smhasher as default hasher and add options for cityhash and hashxx
+
+* Initial version
+---------------------
 0.1.0 (2016-11-16)
 ---------------------